refactor(sdk): consolidate all tobigint usage
